refactor: block watcher committer merger (#83)
