+++ packages/services/Cargo.toml
@@ -16,7 +16,6 @@ metrics = { workspace = true }
 ports = { workspace = true, features = ["full"] }
 serde = { workspace = true }
 thiserror = { workspace = true }
-tokio = { workspace = true, features = ["sync"] }
 tokio-util = { workspace = true }
 tracing = { workspace = true }
 validator = { workspace = true }
@@ -24,7 +23,9 @@ validator = { workspace = true }
 [dev-dependencies]
 fuel-crypto = { workspace = true, features = ["random"] }
 mockall = { workspace = true }
+ports = { workspace = true, features = ["full", "test-helpers"] }
 rand = { workspace = true }
 storage = { workspace = true, features = ["test-helpers"] }
 tai64 = { workspace = true }
+tokio = { workspace = true, features = ["macros"] }
 validator = { workspace = true, features = ["test-helpers"] }
