+++ packages/services/src/lib.rs
@@ -1,13 +1,11 @@
 #![deny(unused_crate_dependencies)]
 mod block_committer;
-mod block_watcher;
 mod commit_listener;
 mod health_reporter;
 mod status_reporter;
 mod wallet_balance_tracker;
 
 pub use block_committer::BlockCommitter;
-pub use block_watcher::BlockWatcher;
 pub use commit_listener::CommitListener;
 pub use health_reporter::HealthReporter;
 pub use status_reporter::StatusReporter;
